RBCFTools: 'BCFTools', 'libbcftools' and 'htslib' Wrappers and 'BCF'/'VCF' to 'Parquet' Convertors

Bundles the 'htslib' and 'bcftools' libraries and command lines tools for reading and manipulating VCF/BCF files. Includes streaming facilities from VCF to Apache Arrow via 'nanoarrow', enabling export to Arrow IPC format and Parquet format using 'duckdb' including a 'bcf_reader' extension. Utilities for reading and writing VCF/BCF files into 'DuckLake' provided.

See also

Author

Maintainer: Sounkou Mahamane Toure sounkoutoure@gmail.com

Other contributors:

Bonfield, James K and Marshall, John and Danecek, Petr and Li, Heng and Ohan, Valeriu and Whitwham, Andrew and Keane, Thomas Davies, Robert M, Pierre Lindenbaum (Authors of included htslib library and bcftools command line tools) [copyright holder]

Zilong Li zilong.dk@gmail.com (Author of the vcfpp library from whom makefiles and configure strategy is borrowed) [copyright holder]

Duckdb C API and extension and API authors (Authors of the duckdb extension and API used for parquet export) [copyright holder]